pyproject.toml
[project]
name = "bank-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "mcp>=1.26.0",
    "pillow>=12.1.1",
    "pyjwt>=2.11.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
